requirements.txt
flask>=2.0.3
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
pymongo[srv]>=4.2.0
setuptools==63.2.0
Flask-pyoidc>=3.11.0
pika>=1.3.1
pyjwt>=2.6.0
Flask-Session>=0.4.0